MOD : fonctionnement des routes par défaut

+++ Web/RDD.Web/Controllers/WebController.cs
@@ -33,44 +33,14 @@ namespace RDD.Web.Controllers
         {
         }
 
-        public Task<IActionResult> PostAsync()
+        [HttpPost]
+        public virtual async Task<IActionResult> PostAsync()
         {
-            if (AllowedHttpVerbs.HasVerb(HttpVerbs.Post))
+            if (!AllowedHttpVerbs.HasVerb(HttpVerbs.Post))
             {
-                return ProtectedPostAsync();
+                return Unauthorized();
             }
-            return Task.FromResult((IActionResult)NotFound());
-        }
 
-        public Task<IActionResult> PutByIdAsync(TKey id)
-        {
-            if (AllowedHttpVerbs.HasVerb(HttpVerbs.Put))
-            {
-                return ProtectedPutAsync(id);
-            }
-            return Task.FromResult((IActionResult)NotFound(id));
-        }
-
-        public Task<IActionResult> PutAsync()
-        {
-            if (AllowedHttpVerbs.HasVerb(HttpVerbs.Put))
-            {
-                return ProtectedPutAsync();
-            }
-            return Task.FromResult((IActionResult)NotFound());
-        }
-
-        public Task<IActionResult> DeleteByIdAsync(TKey id)
-        {
-            if (AllowedHttpVerbs.HasVerb(HttpVerbs.Delete))
-            {
-                return ProtectedDeleteAsync(id);
-            }
-            return Task.FromResult((IActionResult)NotFound(id));
-        }
-
-        protected virtual async Task<IActionResult> ProtectedPostAsync()
-        {
             Query<TEntity> query = Helper.CreateQuery(HttpVerbs.Post, false);
             ICandidate<TEntity, TKey> candidate = Helper.CreateCandidate();
 
@@ -79,8 +49,14 @@ namespace RDD.Web.Controllers
             return Ok(RDDSerializer.Serialize(entity, query));
         }
 
-        protected virtual async Task<IActionResult> ProtectedPutAsync(TKey id)
+        [HttpPut("{id}")]
+        public virtual async Task<IActionResult> PutByIdAsync(TKey id)
         {
+            if (!AllowedByIdHttpVerbs.HasVerb(HttpVerbs.Put))
+            {
+                return Unauthorized();
+            }
+
             Query<TEntity> query = Helper.CreateQuery(HttpVerbs.Put, false);
             ICandidate<TEntity, TKey> candidate = Helper.CreateCandidate();
 
@@ -94,14 +70,20 @@ namespace RDD.Web.Controllers
             return Ok(RDDSerializer.Serialize(entity, query));
         }
 
-        protected virtual async Task<IActionResult> ProtectedPutAsync()
+        [HttpPut]
+        public virtual async Task<IActionResult> PutAsync()
         {
+            if (!AllowedHttpVerbs.HasVerb(HttpVerbs.Put))
+            {
+                return Unauthorized();
+            }
+
             Query<TEntity> query = Helper.CreateQuery(HttpVerbs.Put, false);
             IEnumerable<ICandidate<TEntity, TKey>> candidates = Helper.CreateCandidates();
 
             if (candidates.Any(c => !c.HasId()))
             {
-                return BadRequest("To edit a collection of entities, provide an array of objets with an property id");
+                return BadRequest("To edit a collection of entities, provide an array of objets with an 'id' property");
             }
 
             var candidatesByIds = candidates.ToDictionary(c => c.Id);
@@ -111,26 +93,36 @@ namespace RDD.Web.Controllers
             return Ok(RDDSerializer.Serialize(entities, query));
         }
 
-        protected virtual async Task<IActionResult> ProtectedDeleteAsync(TKey id)
+        [HttpDelete("{id}")]
+        public virtual async Task<IActionResult> DeleteByIdAsync(TKey id)
         {
+            if (!AllowedByIdHttpVerbs.HasVerb(HttpVerbs.Delete))
+            {
+                return Unauthorized();
+            }
+
             await AppController.DeleteByIdAsync(id);
 
             return Ok();
         }
 
-        protected virtual async Task<IActionResult> ProtectedDeleteAsync()
+        [HttpDelete]
+        public virtual async Task<IActionResult> DeleteAsync()
         {
+            if (!AllowedHttpVerbs.HasVerb(HttpVerbs.Delete))
+            {
+                return Unauthorized();
+            }
+
             Query<TEntity> query = Helper.CreateQuery(HttpVerbs.Delete);
             IEnumerable<ICandidate<TEntity, TKey>> candidates = Helper.CreateCandidates();
 
             if (candidates.Any(c => !c.HasId()))
             {
-                return BadRequest("To delete a collection of entities, provide an array of objets with an property id");
+                return BadRequest("To delete a collection of entities, provide an array of objets with an 'id' property");
             }
 
-            var ids = candidates.Select(c => c.Id).ToList();
-
-            await AppController.DeleteByIdsAsync(ids);
+            await AppController.DeleteByIdsAsync(candidates.Select(c => c.Id));
 
             return Ok();
         }
